# magical 1-byte speed increase