Creating & destroying an array of 6 strings can add up over the course of some 19,630,920 iterations. Passing the datacheck FP entries by reference instead of value avoids this, yielding a substantial speed boost. Seems the bottleneck here is memory bandwidth; the more bandwidth my machines have, the more muted the effect of the speedup.

* BiggaTomato (DDR2-800): 10.3s -> 2.6s, a 3.96x speedup.
* Epoch (DDR3-1333): 10.7s -> 4.1s, a 2.61x speedup.
* lab2 (DDR3-1600): 4.1s -> 1.9s, a 2.16x speedup.
* noreaster is the exception: 6.7s -> 1.3s ~= 5.15x. I'll just assume that noreaster has a good number of other processes running and competing for bandwidth. (I *assume* noreaster has the highest bandwidth. It could be running at <1600 mhz, or even single-channel for all I know...)