@yakra 2019-12-12
> Seems the bottleneck here may be memory bandwidth

Or not. I installed a single stick of DDR3-1333 in lab2. 4.2s -> 2.2s. So there are other factors at play here. Cache size, maybe? Epoch has 512K, BiggaTomato 8MB, and lab2 has 512K L2 + 20M L3. And if noreaster has a bunch of other processes running, that could mean more data in the cache for them, and more cache misses for siteupdate. In any case, the  speedup is real, and it's a Good Thing.

> Just missed by a few weeks being in time to show this as an example of differences in calling conventions in my programming languages class.

:D
I've scoured the code looking for opportunities for similar improvements. There are a few, but nothing of this much benefit, that gets repeated this many times.
Other possibilities to check:
* range `for` loops
* return values