@jteresco 2019-12-12
Noreaster has all the web service and DB service, which I suppose could be using some memory bandwidth, but I wouldn't expect that much.  No CPU intensive or other memory intensive jobs are normally running at this point.  The machine's been up for a while so the bootup logs aren't easy to see at the moment to remind myself of its memory size and type and cache sizes.

In any case, a nice speedup!